Moved some code for better organization
